refactor(market-data): Phase5b extract md-sidefx worker module
Extract md-sidefx OS thread (commands, bounded enqueue, burst coalesce,
all md_sidefx_process_* handlers) into src/market_data/sidefx/.

- worker.rs: MdSidefxCommand, MdSidefxSender, spawn_md_sidefx_worker
- handlers.rs: cache + NATS side-effects only (Phase-1: no md-state register)
- host.rs: SidefxWorkerHost trait; bin MarketDataSidefxHost adapter
- pool_publish.rs: JetStream PoolCacheUpdate metadata helpers

Bin keeps thin eval-grep wrappers and impl SidefxWorkerHost.
Phase-1/5a/5b unit tests green; fmt + clippy clean.
