+++ src/market_data/sidefx/mod.rs
@@ -1,0 +1,19 @@
+pub mod handlers;
+pub mod host;
+pub mod pool_publish;
+pub mod worker;
+
+pub use handlers::{
+    md_sidefx_process_bonding_curve, md_sidefx_process_generic_dex_first_trade,
+    md_sidefx_process_job, md_sidefx_process_live_pool_cache_account_update,
+    md_sidefx_process_live_pool_cache_mint_decimals, md_sidefx_process_pump_amm_create_pool,
+    md_sidefx_process_pump_amm_trade, md_sidefx_process_pump_fun_dev_wallet_from_pool_created,
+    md_sidefx_process_pump_fun_pool_mint_map, md_sidefx_process_touch_bin_array_tick,
+    md_sidefx_process_trade_pool_lru_touch, md_sidefx_process_vault_balance_tick,
+};
+pub use host::{MarketEventCorePublishTrace, SidefxVaultMembershipView, SidefxWorkerHost};
+pub use worker::{
+    md_sidefx_coalesce_burst, md_sidefx_coalesce_key, md_sidefx_flush_pending_md_state_jobs,
+    md_sidefx_try_enqueue, spawn_md_sidefx_worker, MdSidefxBurstScratch, MdSidefxCommand,
+    MdSidefxSender, MARKET_DATA_MD_SIDEFX_BURST_MAX, MARKET_DATA_MD_SIDEFX_QUEUE_CAP,
+};
